refactor(#578): dissolve module lib/ subdirs → co-locate at module root
Standardize helper placement: dissolve the `lib/` subdirs in the
analytics, messaging, and partner_billing modules, co-locating the pure
helpers at the module root next to their service. Pure import-path move,
no behavior change.

- modules/analytics/lib/{compute-daily-stats,country-from-timezone,ingest-buffer}.ts → module root
- modules/messaging/lib/record-whatsapp-notification.ts → module root
- modules/partner_billing/lib/{compute-fee,describe-fee,resolve-fee-rate,summarize-fees}.ts → module root
- moved co-located __tests__ up a level; updated all importers + integration spec

workflows/email/lib/ deferred — it needs a *-lib.ts rename + placement
decision (tracked on #578), not a plain path move.

Verify: tsc baseline unchanged (69, 0 new); 166 unit tests pass.
