+++ apps/backend/src/api/admin/ops/maintenance-jobs/registry.ts
@@ -9,8 +9,8 @@ import { RAW_MATERIAL_MODULE } from "../../../../modules/raw_material"
 import { OPS_AUDIT_MODULE } from "../../../../modules/ops_audit"
 import { STATS_MODULE } from "../../../../modules/stats"
 import { PARTNER_BILLING_MODULE } from "../../../../modules/partner_billing"
-import { computeFee } from "../../../../modules/partner_billing/lib/compute-fee"
-import { resolvePartnerFeeRate } from "../../../../modules/partner_billing/lib/resolve-fee-rate"
+import { computeFee } from "../../../../modules/partner_billing/compute-fee"
+import { resolvePartnerFeeRate } from "../../../../modules/partner_billing/resolve-fee-rate"
 import partnerOrderLink from "../../../../links/partner-order"
 import partnerRegionLink from "../../../../links/partner-region"
 import productGoogleMerchantLink from "../../../../links/product-google-merchant-link"
