+++ apps/backend/src/api/partners/[id]/fees/route.ts
@@ -9,7 +9,7 @@ import { AuthenticatedMedusaRequest, MedusaResponse } from "@medusajs/framework"
 import { MedusaError } from "@medusajs/framework/utils"
 import { getPartnerFromAuthContext } from "../../helpers"
 import { PARTNER_BILLING_MODULE } from "../../../../modules/partner_billing"
-import { summarizePartnerFees } from "../../../../modules/partner_billing/lib/summarize-fees"
+import { summarizePartnerFees } from "../../../../modules/partner_billing/summarize-fees"
 
 /**
  * GET /partners/[id]/fees
