+++ apps/backend/src/modules/analytics/compute-daily-stats.ts
@@ -1,4 +1,4 @@
-import AnalyticsService from "../service"
+import AnalyticsService from "./service"
 
 export type DailyStatsRow = {
   website_id: string
